+++ ui/app/reducers/app.js
@@ -23,6 +23,9 @@ function reduceApp(state, action) {
 
   var appState = extend({
     currentView: seedWords ? seedConfView : defaultView,
+    accountDetail: {
+      subview: 'transactions',
+    },
     currentDomain: 'example.com',
     transForward: true, // Used to render transition direction
     isLoading: false,   // Used to display loading indicator
@@ -109,7 +112,9 @@ function reduceApp(state, action) {
   case actions.UNLOCK_METAMASK:
     return extend(appState, {
       currentView: {},
+      detailView: {},
       transForward: true,
+      isLoading: false,
       warning: null,
     })
 
@@ -131,6 +136,7 @@ function reduceApp(state, action) {
     return extend(appState, {
       currentView: {},
       accountDetail: {
+        subview: 'transactions',
         accountExport: 'none',
         privateKey: '',
       },
@@ -144,6 +150,7 @@ function reduceApp(state, action) {
         context: action.value || account,
       },
       accountDetail: {
+        subview: 'transactions',
         accountExport: 'none',
         privateKey: '',
       },
@@ -157,6 +164,7 @@ function reduceApp(state, action) {
         context: action.value,
       },
       accountDetail: {
+        subview: 'transactions',
         accountExport: 'none',
         privateKey: '',
       },
@@ -217,6 +225,9 @@ function reduceApp(state, action) {
         currentView: {
           name: 'accountDetail',
           context: state.metamask.selectedAddress,
+        },
+        accountDetail: {
+          subview: 'transactions',
         },
       })
     }
@@ -285,7 +296,13 @@ function reduceApp(state, action) {
 
   case actions.REQUEST_ACCOUNT_EXPORT:
     return extend(appState, {
+      transForward: true,
+      currentView: {
+        name: 'accountDetail',
+        context: appState.currentView.context,
+      },
       accountDetail: {
+        subview: 'export',
         accountExport: 'requested',
       },
     })
@@ -293,6 +310,7 @@ function reduceApp(state, action) {
   case  actions.EXPORT_ACCOUNT:
     return extend(appState, {
       accountDetail: {
+        subview: 'export',
         accountExport: 'completed',
       },
     })
@@ -300,6 +318,7 @@ function reduceApp(state, action) {
   case  actions.SHOW_PRIVATE_KEY:
     return extend(appState, {
       accountDetail: {
+        subview: 'export',
         accountExport: 'completed',
         privateKey: action.value,
       },